+++ src/main/java/apiquality/sonar/asyncapi/checks/RulesLists.java
@@ -34,7 +34,8 @@ public final class RulesLists {
             AAR036BadDescriptionCheck.class,
             AAR037BindingVersionCheck.class,
             AAR042MessageIdentifierCheck.class,
-            AAR050InfoTitleRequiredCheck.class
+            AAR050InfoTitleRequiredCheck.class,
+            AAR051OperationIdCamelCaseCheck.class
         );
     }
 
@@ -72,7 +73,8 @@ public final class RulesLists {
             AAR046AvroRecordDocCheck.class,
             AAR047AvroFieldDocCheck.class,
             AAR048AvroNameNomenclatureCheck.class,
-            AAR049AvroDefaultNullCheck.class
+            AAR049AvroDefaultNullCheck.class,
+            AAR052AvroNamespacePatternCheck.class
             );
     }